# clients-data-service-test-app/src/main/java/ru/softlab/zikova/efr/controller/ProductsApi.java
/*
 * clients-data-service-test
 * Тестовый API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package ru.softlab.zikova.efr.controller;

import ru.softlab.zikova.efr.exchange.model.Product;
import ru.softlab.zikova.efr.exchange.model.PublicModelFormatter;
import org.springframework.core.annotation.Order;
import org.springframework.core.Ordered;
import ru.softlab.efr.infrastructure.logging.api.model.OperationLogEntry;
import ru.softlab.efr.infrastructure.logging.api.model.OperationMode;
import ru.softlab.efr.infrastructure.logging.api.model.OperationState;
import ru.softlab.efr.infrastructure.logging.api.services.OperationLogService;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.*;
import org.springframework.validation.*;
import org.springframework.data.web.*;
import org.springframework.data.domain.*;

import java.util.*;

import org.springframework.validation.annotation.*;
import javax.validation.constraints.*;
import javax.validation.*;

/**
 * Содержит эндпоинты приложения 'Тестовый API'
 * Продукты
 */
@SuppressWarnings("all")
@RestController
public class ProductsApi {
    public static final String DELETE_PRODUCT_PATH = "/product/{id}";
    public static final String FIND_PRODUCT_BY_ID_PATH = "/product/{id}";
    public static final String SAVE_PRODUCT_PATH = "/product";
    public static final String UPDATE_PRODUCT_PATH = "/product";

    private static final long LOG_TYPE_ID = 10;

    private final Delegate delegate;
    private final OperationLogService operationLogService;

    /**
     * Конструктор
     *
     * @param delegate эффективная реализации контроллера
     * @param operationLogService сервис логирования
     */
    public ProductsApi(Delegate delegate, OperationLogService operationLogService) {
        this.delegate = delegate;
        this.operationLogService = operationLogService;
    }

    /**
     * Удаление данных продукта
     *
     * @param id Идентификатор продукта
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = DELETE_PRODUCT_PATH,
            method = RequestMethod.DELETE)
    public ResponseEntity<Void> deleteProduct(@PathVariable("id") Long id) throws Exception {
        OperationLogEntry operationLogEntry = operationLogService.startLogging(LOG_TYPE_ID);
        operationLogEntry.setLogTypeName("Обработка запроса сервисом " + operationLogEntry.getLoggingSystem());
        operationLogEntry.setOperationKey(DELETE_PRODUCT_PATH);
        operationLogEntry.setOperationDescription("Удаление данных продукта");
        operationLogEntry.setOperationMode(OperationMode.ACTIVE);
        operationLogEntry.setOperationParameter("Идентификатор продукта", PublicModelFormatter.format(id));
        operationLogEntry.setSubjectId(PublicModelFormatter.format(id));
        try {
            ResponseEntity<Void> responseEntity = delegate.deleteProduct(id);
            operationLogEntry.setOperationState(getOperationState(responseEntity));
            return responseEntity;
        } catch (Exception e) {
            operationLogEntry.setOperationState(OperationState.SYSTEM_ERROR);
            throw e;
        } finally {
            writeLog(operationLogEntry);
        }
    }

    /**
     * Получение продукта по id
     *
     * @param id Идентификатор продукта
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = FIND_PRODUCT_BY_ID_PATH,
            produces = { "application/json" }, 
            method = RequestMethod.GET)
    public ResponseEntity<Product> findProductById(@PathVariable("id") Long id) throws Exception {
        OperationLogEntry operationLogEntry = operationLogService.startLogging(LOG_TYPE_ID);
        operationLogEntry.setLogTypeName("Обработка запроса сервисом " + operationLogEntry.getLoggingSystem());
        operationLogEntry.setOperationKey(FIND_PRODUCT_BY_ID_PATH);
        operationLogEntry.setOperationDescription("Получение продукта по id");
        operationLogEntry.setOperationMode(OperationMode.PASSIVE);
        operationLogEntry.setOperationParameter("Идентификатор продукта", PublicModelFormatter.format(id));
        operationLogEntry.setSubjectId(PublicModelFormatter.format(id));
        try {
            ResponseEntity<Product> responseEntity = delegate.findProductById(id);
            operationLogEntry.setOperationState(getOperationState(responseEntity));
            return responseEntity;
        } catch (Exception e) {
            operationLogEntry.setOperationState(OperationState.SYSTEM_ERROR);
            throw e;
        } finally {
            writeLog(operationLogEntry);
        }
    }

    /**
     * Добавление продукта в БД
     *
     * @param product Данные продукта
     * @param bindingResult информация о валидности входящих данных
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = SAVE_PRODUCT_PATH,
            method = RequestMethod.POST)
    public ResponseEntity<Product> saveProduct(@Valid @RequestBody Product product, BindingResult bindingResult) throws Exception {
        OperationLogEntry operationLogEntry = operationLogService.startLogging(LOG_TYPE_ID);
        operationLogEntry.setLogTypeName("Обработка запроса сервисом " + operationLogEntry.getLoggingSystem());
        operationLogEntry.setOperationKey(SAVE_PRODUCT_PATH);
        operationLogEntry.setOperationDescription("Добавление продукта в БД");
        operationLogEntry.setOperationMode(OperationMode.ACTIVE);
        operationLogEntry.setOperationParameter("Данные продукта", PublicModelFormatter.format(product));
        if (bindingResult.hasErrors()) {
            operationLogEntry.setOperationState(OperationState.CLIENT_ERROR);
            writeLog(operationLogEntry);
            throw new BindException(bindingResult);
        }
        try {
            ResponseEntity<Product> responseEntity = delegate.saveProduct(product);
            operationLogEntry.setOperationState(getOperationState(responseEntity));
            return responseEntity;
        } catch (Exception e) {
            operationLogEntry.setOperationState(OperationState.SYSTEM_ERROR);
            throw e;
        } finally {
            writeLog(operationLogEntry);
        }
    }

    /**
     * Обновление данных продукта
     *
     * @param product Данные продукта
     * @param bindingResult информация о валидности входящих данных
     * @return Экземпляр класса {@link ResponseEntity}
     * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
     */
    @RequestMapping(value = UPDATE_PRODUCT_PATH,
            method = RequestMethod.PUT)
    public ResponseEntity<Product> updateProduct(@Valid @RequestBody Product product, BindingResult bindingResult) throws Exception {
        OperationLogEntry operationLogEntry = operationLogService.startLogging(LOG_TYPE_ID);
        operationLogEntry.setLogTypeName("Обработка запроса сервисом " + operationLogEntry.getLoggingSystem());
        operationLogEntry.setOperationKey(UPDATE_PRODUCT_PATH);
        operationLogEntry.setOperationDescription("Обновление данных продукта");
        operationLogEntry.setOperationMode(OperationMode.ACTIVE);
        operationLogEntry.setOperationParameter("Данные продукта", PublicModelFormatter.format(product));
        if (bindingResult.hasErrors()) {
            operationLogEntry.setOperationState(OperationState.CLIENT_ERROR);
            writeLog(operationLogEntry);
            throw new BindException(bindingResult);
        }
        try {
            ResponseEntity<Product> responseEntity = delegate.updateProduct(product);
            operationLogEntry.setOperationState(getOperationState(responseEntity));
            return responseEntity;
        } catch (Exception e) {
            operationLogEntry.setOperationState(OperationState.SYSTEM_ERROR);
            throw e;
        } finally {
            writeLog(operationLogEntry);
        }
    }


    private static OperationState getOperationState(ResponseEntity<?> stringResponseEntity) {
        if (stringResponseEntity.getStatusCode().is4xxClientError()) {
            return OperationState.CLIENT_ERROR;
        }
        if (stringResponseEntity.getStatusCode().is5xxServerError()) {
            return OperationState.SYSTEM_ERROR;
        }
        return OperationState.SUCCESS;
    }

    private void writeLog(OperationLogEntry operationLogEntry) {
        operationLogEntry.setDuration(Calendar.getInstance().getTimeInMillis() - operationLogEntry.getLogDate().getTimeInMillis());
        operationLogService.log(operationLogEntry);
    }

    /**
     * Интерфейс эффективной реализации контроллера
     */
    @Order(Ordered.HIGHEST_PRECEDENCE)
    @RestControllerAdvice(assignableTypes = ProductsApi.class)
    public interface Delegate {
        /**
         * Удаление данных продукта
         *
         * @param id Идентификатор продукта 
         * @return Экземпляр класса {@link ResponseEntity}
         * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
         */
        ResponseEntity<Void> deleteProduct(Long id) throws Exception;

        /**
         * Получение продукта по id
         *
         * @param id Идентификатор продукта 
         * @return Экземпляр класса {@link ResponseEntity}
         * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
         */
        ResponseEntity<Product> findProductById(Long id) throws Exception;

        /**
         * Добавление продукта в БД
         *
         * @param product Данные продукта 
         * @return Экземпляр класса {@link ResponseEntity}
         * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
         */
        ResponseEntity<Product> saveProduct(Product product) throws Exception;

        /**
         * Обновление данных продукта
         *
         * @param product Данные продукта 
         * @return Экземпляр класса {@link ResponseEntity}
         * @throws Exception исключение, выбрасываемое в случае ошибок при выполнении метода
         */
        ResponseEntity<Product> updateProduct(Product product) throws Exception;
    }
}